add manage product of admin

+++ src/main/java/com/tthieu/controller/admin/api/CategoryAPI.java
@@ -47,9 +47,12 @@ public class CategoryAPI extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
         req.setCharacterEncoding("UTF-8");
-        int id  = Integer.parseInt(req.getParameter("id"));
-        System.out.println(id);
-        categoryService.delete(id);
-        resp.sendRedirect(req.getContextPath() + "/admin-category");
+        if (req.getParameter("action").equals("delete")){
+            int id  = Integer.parseInt(req.getParameter("id"));
+            System.out.println(id);
+            categoryService.delete(id);
+            resp.sendRedirect(req.getContextPath() + "/admin-category");
+        }
+
     }
 }
